perms

+++ 2010/3/different-sum/src/main/java/com/eric/codejam/InputData.java
@@ -4,7 +4,7 @@ import com.eric.codejam.main.AbstractInputData;
 
 public class InputData extends AbstractInputData {
     //Données
-    int N;
+    long N;
     int B;
     
     InputData(int testCase) {
